pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pm-bot"
version = "0.0.0"
description = "Agent-native PM orchestrator (scaffold)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "ruff>=0.3.0",
]

[project.scripts]
pm = "pm_bot.cli:app"

[tool.setuptools.packages.find]
include = ["pm_bot", "pm_bot.*"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"
